+++ src/sagemaker/session.py
@@ -202,7 +202,7 @@ class Session(object):
         return self._default_bucket
 
     def train(self, image, input_mode, input_config, role, job_name, output_config,
-              resource_config, hyperparameters, stop_condition, tags):
+              resource_config, vpc_config, hyperparameters, stop_condition, tags):
         """Create an Amazon SageMaker training job.
 
         Args:
@@ -258,6 +258,9 @@ class Session(object):
 
         if tags is not None:
             train_request['Tags'] = tags
+
+        if vpc_config is not None:
+            train_request['VpcConfig'] = vpc_config
 
         LOGGER.info('Creating training-job with name: {}'.format(job_name))
         LOGGER.debug('train request: {}'.format(json.dumps(train_request, indent=4)))